pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "socrates-ai"
version = "0.1.0"
description = "Rubric-guided content generation for publishable drafts."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "miounet11" },
]
keywords = [
  "ai",
  "content-generation",
  "llm",
  "marketing",
  "prompt-engineering",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Text Processing",
]
dependencies = [
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "PyYAML>=6.0.2,<7.0.0",
  "typer>=0.15.2,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.3.5,<9.0.0",
  "pytest-cov>=6.0.0,<7.0.0",
  "ruff>=0.11.0,<1.0.0",
  "types-PyYAML>=6.0.12.20250516,<7.0.0",
]

[project.scripts]
socrates = "socrates.cli:app"

[project.urls]
Homepage = "https://github.com/miounet11/Socrates"
Repository = "https://github.com/miounet11/Socrates"
Issues = "https://github.com/miounet11/Socrates/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/socrates"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "B",
  "E",
  "F",
  "I",
  "N",
  "RUF",
  "UP",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
plugins = []
